Add listener helper class for battle resolution test
